fix(coverage): route transit around obstacle holes instead of through them (#333)
A forward turn-around connector is a LOCAL Dubins arc — it cannot route around
a large interior obstacle. The connectors that join rings/swaths into the one
continuous path were validated only against the outer boundary, so between two
hole-split segments the join fell back to a straight line straight THROUGH the
hole (measured 959/9062 path poses inside a central hole). Safety: the blade
crossed forbidden obstacle zones during transitions.

Fix: split the continuous path into hole-free continuous SUB-PATHS wherever a
connector can't avoid a hole, and let the BT bridge the gap the way it already
bridges any >0.6 m segment gap — a blade-off, costmap/hole-aware Nav2
NavigateToPose transit that routes AROUND the obstacle.

- coverage_planning: planBoustrophedon exposes the inset ('grown') hole rings
  (plan.safe_holes); connectors + corner fillets now validate against them
  (clearOfHoles); buildContinuousSubPaths breaks the path at un-avoidable
  hole-crossings. buildContinuousPath stays the concatenation (GUI full_path).
- coverage_server: fills PlanCoverage.action's new drivable_subpaths (one entry
  for a hole-free field); logs residual in-hole/out-of-bounds poses as errors.
- FollowStrip: drives the sub-paths (one FollowCoveragePath goal each) with the
  existing inter-segment Nav2 transit between them; resume cursor generalised to
  an index into the sub-path concatenation (base offsets), progress reset per
  unit. Single-sub-path (hole-free) fields behave exactly as before.
- test: ContinuousPathAvoidsHole asserts a central hole splits into >=2 hole-free
  sub-paths with no pose inside the hole.

Full split-path + Nav2 transit (not report-only) per maintainer decision.

+++ ros2/src/mowgli_behavior/src/coverage_persistence.cpp
@@ -65,9 +65,9 @@ bool saveCoverageResumeState(const BTContext& ctx)
     if (auto it = ctx.area_path_pose_count.find(idx); it != ctx.area_path_pose_count.end())
       pose_count = it->second;
     // -1 sentinel = no live resume cursor (area finished or never interrupted).
-    long long resume = -1;
+    int64_t resume = -1;
     if (auto it = ctx.area_resume_pose_index.find(idx); it != ctx.area_resume_pose_index.end())
-      resume = static_cast<long long>(it->second);
+      resume = static_cast<int64_t>(it->second);
 
     out << "area " << idx << ' ' << pose_count << ' ' << resume << " completed";
     if (auto it = ctx.area_completed_swaths.find(idx); it != ctx.area_completed_swaths.end())
@@ -138,7 +138,7 @@ bool loadCoverageResumeState(BTContext& ctx)
     {
       uint32_t idx;
       std::size_t pose_count;
-      long long resume;
+      int64_t resume;
       std::string completed_tag;
       if (!(ls >> idx >> pose_count >> resume >> completed_tag))
       {
